// src/app/modules/logistica/estoque/painel-inventario/models/totalInventario.ts
export class LogisticaEstoquePainelInventarioTotalInventario {
  qtNotasFiscais: number;
  qtNotasFiscaisRo: number;
  totalContagem: number;
  saldoInventario: number;
  resultadoInventario: number;
  resultadoMonetario: number;
  enderecoPatio: string;
  qtPesoEstoque: number;
  saldoEstoque: number;
  somaInventarioRotativo: number;
  saldoFinanceiro: number;
  precoMinimo: number;
  pesoPeca: number;
  dtInclusao: string;
  dsClasse: string;
  qtMaterialNotaFiscal: number;
  qtMaterialNotaFiscalRo: number;
}